Remove section 6.5 Types of function call
This section is not quite correct because python only uses call by object
reference and it perhaps goes beyond the scope of beginner introduction to go
through this concept in proper detail.

Python doesn't work in a similar fashion to C (call by value/reference) and it
this may just create more confusion.